Implement faster VA-API based decoder
The main bottleneck of the VA-API decoder is the transfer of the HW
frame into system memory. This change implements 2 better methods to
perform this transfer. The original method is left in place as a 3rd
option, as fallback.

1. Map frame to DRM prime fd

This provides direct access to the raw GPU buffer. For Intel GPU, the
buffer is not linear, but Y-tiled. This requires additional support in
minigbm to support the format with the correct modifiers.

2. Map frame to YV12

This maps the HW frame to memory, using linear YV12 format. Because the
Intel GPU buffer is Y-tiled internally, this requires converting the
tiled content to linear one, which slows down the process.

3. Download frame as YV12

This is the same as 2, but also copy the mapped buffer into FFMPEG
internal buffer. This additional copy introduces further latency.

To ensure backward compatibility, the OMX codec does not use the first
method to transfer HW frames.